praxis_agent/requirements.txt
# Core dependencies
requests>=2.31.0
pystray>=0.19.4
pillow>=10.0.0
pywin32>=306
psutil>=5.9.5
rich>=13.7.0
typer>=0.9.0
pydantic>=2.5.0

# AI and ML dependencies
openai>=1.6.0
anthropic>=0.8.0
ollama>=0.1.0
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2

# Data and storage
chromadb>=0.4.18
json5>=0.9.14
pyyaml>=6.0.1

# Web and automation
selenium>=4.15.0
beautifulsoup4>=4.12.2
playwright>=1.40.0

# Development and testing
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
